Move certs utilities back to utils pkg
