+++ pkg/controller/manager/manager_controller.go
@@ -247,7 +247,7 @@ func (r *ReconcileManager) Reconcile(ctx context.Context, request reconcile.Requ
 	svcDNSNames := dns.GetServiceDNSNames(render.ManagerServiceName, render.ManagerNamespace, r.clusterDomain)
 	svcDNSNames = append(svcDNSNames, "localhost")
 	certDur := 825 * 24 * time.Hour // 825days*24hours: Create cert with a max expiration that macOS 10.15 will accept
-	tlsSecret, err = render.EnsureCertificateSecret(
+	tlsSecret, err = utils.EnsureCertificateSecret(
 		ctx, render.ManagerTLSSecretName, tlsSecret, render.ManagerSecretKeyName, render.ManagerSecretCertName, certDur, instance.GetUID(), svcDNSNames...,
 	)
